+++ NEWS.md
@@ -11,6 +11,12 @@ I am successful in that undertaking.
 Consult [the Github repository](github.com/verumsolum/orf_weather) if you
 require details about earlier changes.
 
+### Changes in version 0.0.0.9041 (2017-01-18)
+* **firstSunday**: Function to return the date of the first Sunday in a month
+* **plotMTDPrecipitation**: Label axis with ticks and labels each Sunday
+* **plotYTDPrecipitation**: Label axis with ticks and labels each Sunday, tweak
+alignment and placement of year labels
+
 ### Changes in version 0.0.0.9040 (2017-01-14)
 * **plotYTDPrecipitation**: New year-to-date plot based on 
 `plotMTDPrecipitation`.